Adds database backup and sync functionality
Implements scheduled production database synchronization to the development environment.

This change introduces a new Netlify function to automate the process of copying data from the production database to the development database on a daily schedule.
It also adds backup capabilities, allowing to save data to a separate backup database.

The database module and associated controller and service are also added.